backend-api/pyproject.toml
[project]
name = "pythonmoneytracker-backend"
version = "1.0.0"
description = "Portfolio AI App Backend API"
requires-python = ">=3.11"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.12.0",
    "bcrypt==5.0.0",
    "blinker==1.9.0",
    "cachetools==6.2.2",
    "certifi==2025.11.12",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "colorama==0.4.6",
    "Flask==3.1.2",
    "Flask-Bcrypt==1.0.1",
    "flask-cors==6.0.1",
    "Flask-SQLAlchemy==3.1.1",
    "google-auth==2.43.0",
    "google-generativeai==0.8.3",
    "greenlet==3.2.4",
    "gunicorn==23.0.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.11",
    "itsdangerous==2.2.0",
    "Jinja2==3.1.6",
    "MarkupSafe==3.0.3",
    "packaging==25.0",
    "Pillow==11.1.0",
    "psycopg2-binary==2.9.11",
    "pyasn1==0.6.1",
    "pyasn1_modules==0.4.2",
    "pydantic==2.12.5",
    "pydantic_core==2.41.5",
    "PyJWT==2.10.1",
    "python-dotenv==1.2.1",
    "requests==2.32.5",
    "rsa==4.9.1",
    "SQLAlchemy==2.0.44",
    "tenacity==9.1.2",
    "typing-inspection==0.4.2",
    "typing_extensions==4.15.0",
    "urllib3==2.5.0",
    "waitress==3.0.2",
    "websockets==15.0.1",
    "Werkzeug==3.1.4",
    "flask-swagger-ui==4.11.1",
    "pymongo==4.10.1",
    "APScheduler==3.10.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
